requirements.txt
# GooooooAI Talk - Football Match Q&A bot
# Last Updated: 2025-03-13

# HTTP Requests Library
# Used to fetch football match data from API-Football
requests==2.32.1

# Environment Variables Manager
# Loads sensitive information like API keys from .env files
python-dotenv==1.0.0

# OpenAI API Client
# Handles natural language Q&A using GPT models
openai>=1.0.0,<2.0.0

# Terminal UI enhancement
# Provides beautiful console output for CLI interface
rich==13.7.0

# Web Framework
# Implements web interface API endpoints
flask==2.0.1

# HTML Template Engine
# Renders HTML pages for web interface
jinja2==3.1.3

# JSON handling library
jsonlib==1.6.1